requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
defusedxml==0.7.1
distro==1.9.0
Django==4.1.5
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.3
httpie==3.2.2
httpx==0.26.0
idna==3.6
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
openai==1.12.0
packaging==23.2
psycopg2-binary==2.9.9
pydantic==2.6.1
pydantic_core==2.16.2
Pygments==2.17.2
PyJWT==2.8.0
PySocks==1.7.1
pytz==2024.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.0
setuptools==69.1.0
sniffio==1.3.0
sqlparse==0.4.4
tqdm==4.66.2
typing_extensions==4.9.0
urllib3==2.2.0
whitenoise==6.6.0
